Move array hashing into a shared _combine_common substrate
Lifts the array-hashing machinery out of _concatenate.py into a new private
module, lib/iris/_combine_common.py, so that _merge.py can reach it without
importing _concatenate.py.  The substrate imports nothing from Iris at
runtime; the coordinate annotations on array_id sit behind TYPE_CHECKING.

Names lose their underscore prefix on the move, the module itself already
being private.  The two local variables that would otherwise have shadowed
the imported array_id are renamed to points_id and bounds_id.

Pure relocation: no behaviour change.  Row 1 of the roadmap in
docs/src/developers_guide/specs/2026-09-10-merge-concatenate-design.md.
