+++ lib/iris/tests/unit/combine_common/__init__.py
@@ -1,0 +1,5 @@
+# Copyright Iris contributors
+#
+# This file is part of Iris and is released under the BSD license.
+# See LICENSE in the root of the repository for full licensing details.
+"""Unit tests for the :mod:`iris._combine_common` module."""
